+++ team-builder/src/Components/Form.js
@@ -2,31 +2,19 @@ import React, { useState } from 'react';
 import MemberList from './MemberList';
 
 
-
 function Form (props){
 
-  const [teamMembers, setTeamMembers] = useState([])
-      
-  const handleSubmit = e => {
-    e.preventDefault();
-    const newMember = {
-      name: e.target.name.value,
-      email: e.target.email.value,
-      role: e.target.role.value
-    }
-    setTeamMembers(teamMembers.concat(newMember))
-    e.target.reset();
-  }
 
         return (
             <div>
                 <h1>Team Members Form</h1>
-                <form onSubmit={handleSubmit}>
+                <form onSubmit={props.handleSubmit}>
                     <label>Name: 
                         <input 
                         type="text"
                         name="name"
                         placeholder="Enter your name here" 
+                        onChange={props.handleChange}
                        /> 
                     </label>
                     <label>email: 
@@ -34,12 +22,14 @@ function Form (props){
                         type="email"
                         name="email"
                         placeholder="Enter your email here" 
+                        onChange={props.handleChange}
                         /> 
                     </label>
                     <label>Role: 
                       <select
                       type="text"
                       name="role"
+                      onChange={props.handleChange}
                       >
                         <option></option>
                         <option>UI</option>
@@ -50,21 +40,7 @@ function Form (props){
                       </select>
                     </label>
                     <input type="submit" />
-                </form>
-                <h2>Team Members</h2>
-                { teamMembers.length === 0 ? 
-                      <div>
-                        <p>There are no members in this team.</p>
-                      </div> :
-                  teamMembers.map((curr, index) => {
-                    return (
-                        <MemberList 
-                          key={index}
-                          curr={curr}/>
-                          ) 
-                        }) 
-                }
-                
+                </form>            
       </div>
         )
  }
