+++ team-builder/src/Components/MemberList.js
@@ -1,17 +1,21 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+
 
 function MemberList (props) {
 
-        return (
-            <div>
-                            <div key={props.index}>
-                                <h4>Name: {props.curr.name}</h4>
-                                <p>Email: {props.curr.email}</p>
-                                <p>Role: {props.curr.role}</p>
-                            </div>
 
-            </div>
-        )
+
+    return (
+        <div>
+                        <div key={props.index}>
+                            <h4>Name: {props.curr.name}</h4>
+                            <p>Email: {props.curr.email}</p>
+                            <p>Role: {props.curr.role}</p>
+                        </div>
+                        <button >Edit</button>
+
+        </div>
+    )
     
 }
